GPO: Prepare core engine for GPO file portion evaluation
According to [MS-GPOL] group policy clients must also download and evaluate
the file portion of a group policy object, that is stored on SYSVOL, during
policy application.

Current SSSD implementation does download the GPT.INI file of a GPO but very
late as part of the CSE specific policy application.

For subsequent patches, that fix the GPO version evaluation, and for better
compliance to [MS-GPOL] this patch adds a GPO filtering event to the group
policy core engine. The core engine now incorporates the following general
policy application events that are not client-side extensions:
- DC discovery and AD connection establishment ([MS-GPOL] 3.2.5.1.1)
- Process policy target ([MS-GPOL] 3.2.5.1.2)
- Process SOM for domain SOM and site search ([MS-GPOL] 3.2.5.1.3 f.)
- GPO search for search and retrieval of candidate GPOs and their attributes
  ([MS-GPOL] 3.2.5.1.4)
- GPO filtering ([MS-GPOL] 3.2.5.1.5).

If the calling function passes the CSE GUID of a required client-side
extension to the group policy core engine, the core engine will return a
subset of the filtered GPO list, that includes only those filtered GPOs whose
extension names attributes contain the requested CSE GUID ([MS-GPOL]
3.2.5.1.10).

Signed-off-by: REIM THOMAS <reimth@gmail.com>