GPO: Deny empty GPOs during GPO filtering
Current SSSD GPO implementation does only retrieve and evaluate the file system
GPO version, which is stored in the GPT.INI policy file on SYSVOL.

[MS-GPOL] requires group policy clients to retrieve also GPO attribute
versionNumber which holds the container version (i.e. AD version) of a group
policy object. SYSVOL and AD version attributes contain both, a user policy and
a computer policy version. The former is stored in the upper 16 bits and the
latter in the lower 16 bits of the 32 bit version attributes.

Depending on the policy application mode (user or computer) the group policy
client must check either user or the computer part of the version attributes.
Current SSSD implementation performs its checks on the complete 32 bit version
value, which is not correct and might lead to unnecessary network traffic.

This patch implements the following changes:
- Retrieve AD GPO attribute versionNumber during GPO search
- Store AD and SYSVOL GPO version in GPO cache
- Deny GPOs that have both AD and SYSVOL version value set to 0 (GPO
  filtering).

Signed-off-by: REIM THOMAS <reimth@gmail.com>