GPO: GPO flags evaluation not in line with [MS-GOPL]
Current SSSD implementation retrieves the GPO flags from AD and checks for a
GPO flags value = 2 (computer portion disabled) to deny the GPO.

[MS-GOPL] 3.2.5.1.6 requests group policy clients to only check the two least
significant bits of the GPO flags attribute value. If bit 0 is set, the user
portion of the GPO is disabled. If bit 1 is set, the computer portion of the
GPO is disabled.

Signed-off-by: REIM THOMAS <reimth@gmail.com>