GPO: Use cached GPOs if group policy server is offline
Current SSSD implementation performs offline GPO operation if connection
to the AD domain controller fails. The file portion of group policy
objects is stored on SYSVOL, which may be located on a group policy
server different from the domain controller. Current SSSD implementation
will fail if this policy file server is offline and GPO based access
control will not be performed.

This patch identifies a SMB connection timeout to the group policy
server. If the SYSVOL share is offline, SSSD will perform access control
using the cached GPOs in the same way as for for an offline domain
controller.

Signed-off-by: REIM THOMAS <reimth@gmail.com>