GPO: Introduce group policy application core engine
Current SSSD group policy application still does not comply to [MS-GPOL].
The underlying reason for this non-compliance is a simplified, straight forward
implementation of the rather complex [MS-GPOL] standard. The required group
policy application mode was not explicitly defined and stored and terms and
definitions within the code are not comprehensive and often differ from
[MS-GPOL]. This has been corrected by the preceding patch.

This patch provides the following changes:
- Enclose following general policy application events that are not client-side
  extensions by a group policy core engine event handler:
  + DC discovery and AD connection establishment ([MS-GPOL] 3.2.5.1.1)
  + Process policy target (MS-GPOL] 3.2.5.1.2)
  + Process SOM for domain SOM and site search ([MS-GPOL] 3.2.5.1.3 f.)
  + Process GPO for GPO search, retrieval and filter evaluation ([MS-GPOL]
    3.2.5.1.4 and 3.2.5.1.5)
- Simplify the starting access event to trigger following policy application
  tasks:
  + General group policy application (performed by GPO core engine event)
  + Security CSE specific policy application (performed by CSE security task).

This rather minor update clearly isolates policy application for the many and
various possible client side group policy extensions from the single general
group policy application, that is required for all GPOs. In addition, it makes
it easier for SSSD AD users and developers to get along with code and log
messages using [MS-GPOL] as background information.

Signed-off-by: Thomas Reim <reimth@gmail.com>