GPO: Align names and parameters to [MS-GPOL] specification
Current SSSD AD GPO implementation is a straight forward approach for GPO based
access control of users when they logon to the policy client. For this purpose
SSSD uses GPO policy settings that are received from AD. The policy application
operates in computer mode and uses the security client side extension with GUID
{827D319E-6EAC-11D2-A4EA-00C04F79F83A}.

The related rather complex [MS-GPOL] specification to which this a policy
application implementation must comply has been reduced to an absolute minimum
required for SSSD's GPO access control.

Many policy application parameters and settings have not been implemented and
are often implicitily presumed. Variable and function names as well as
explanatory text also follow this simplified approach.

This is the underlying reason for the SSSD GPO client not being compliant to
[MS-GPOL] and behaving differently from Windows workstations and servers. The
related bugs will be fixed by subsequent patches.

This patch prepares the SSSD GPO client for the subsequent patches. The
following changes are provided:
- Introduce policy application mode parameter (user or computer mode)
- Differentiate between logon user and policy target, which are the same
  only in user policy mode
- Clearly outline the specific CSE (security) that is applied during policy
  application
- Remove extra host SID control, as the policy target of access control GPOs is
  the host
- Fix minor code inconsistencies.

This patch does not fix the bugs, but provides a better basis for implementing
the changes and adaptations in a clear, structured and comprehensible way based
on [MS-GPOL].

Signed-off-by: Thomas Reim <reimth@gmail.com>