GPO: Split group policy file download to GPO cache
Current SSSD GPO version control is still not correct. In order to prepare the
implementation for fixing the version control, we split current combined
download of the GPT.INI and the CSE specific policy files.

Download of the GPT.INI file and parsing of the contained file system version
of the GPO is moved to the GPO filtering event within group policy core engine.
Download of the CSE specific policy file (GptTmpl.inf) is continued to be
performed by the CSE specific policy application event.

The GPO child process is simplified to just download the requested files using
the SMB protocol.

Signed-off-by: Thomas Reim <reimth@gmail.com>