GPO: Prepare cache for support of group policy links
Current computer object in SSSD cache of the host/client, for which GPO based
access control is performed, stores the computer name, DN, SID, and the groups
it is member of. For GPO based access control also gpLinks are required, which
store the links to all applicable, security filtered access control GPOs for
that computer.

Following features are implemented:
- Search for a particular client-side extension (CSE) in all cached GPOs
- Remove particular CSEs and their parent GPO from cache:
  Note: GPO cache does not yet implement a timeout for GPO and child CSEs.
        Only a policy file refresh timeout is implemented.
- Store original DN of a GPO
- Store gPLink attribute value(s) in computer object

Signed-off-by: Thomas Reim <reimth@gmail.com>