GPO: Store security CSE settings in GPO cache
In general, a group policy object (GPO) can have a user and computer
part that store the settings of one or more client-side extensions
(CSE). Current SSSD implementation evaluates only the security CSE
specific settings of GPO which are stored in the computer part of the
group policy.

This patch stores and maintaines the container and the file system
version of security CSE group policy objects in the GPO cache. The
security CSE specific group policy file is only retrieved from the group
policy server, if the GPO cache does not contain the GPO or if the
container or file system part of the machine GPO version has changed
([MS-GPOL] 3.2.1.1). In addition, it provides for a periodic refresh of
the group policy file and version information according to [MS-GPOL]
3.2.1.17. Timeout is 90 minutes plus a random offset value in the range
of 0 to 30 minutes. After expiration of the timer, GPO refresh is being
performed as part of the next GPO-based access control, e.g. for login
of a user.

Signed-off-by: Thomas Reim <reimth@gmail.com>